chore(cluster-init): refactor syncrovergroup code
